Rename instance.ts -> raw-instance.ts
