+++ bin/wasm-node/javascript/src/worker/raw-instance.ts
